requirements_backend.txt
flask==3.0.3
mysql-connector-python==9.5.0
requests==2.32.5
numpy==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
threadpoolctl==3.6.0
urllib3==2.5.0
flask-jwt-extended==4.6.0
